Scripts for making coordinates
coordinate_stim.py -> it creates an updated encoding table from an input, that is an encoding table with old coordinates. (it needs an old stimtable excel, e. g. stim_table_enc.xlsx)

stim_practice.py -> it creates practice stimuli with random coordinates, based on the previous practice files, that function as a structure template (it needs the two template xlsx's)